serverless_backend/Module-2User Authentication/GoogleCloudFunctions/requirements.txt
requests==2.18.4
google-auth==1.30.1
google-cloud-firestore==2.0.2
google-auth-oauthlib==0.4.1